Updated fuzzer with --evil parameter
